refactor(app): refactor useSortedProtocols
use useMemo to make the hook efficient and stop using destructive method

close AUTH-
