+++ app/src/resources/protocols/hooks/index.ts
@@ -5,3 +5,4 @@ export * from './useProtocolMetadata'
 export * from './useRequiredProtocolHardware'
 export * from './useRequiredProtocolLabware'
 export * from './useRunTimeParameters'
+export * from './useSortedProtocols'
